+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added `auth login` for registering with `OAuth`
 - Allow user to choose `OAuth` or `Developer Token` on first run
+- Prompt user to `auth login` when API returns unauthorized
 
 ## 2025-05-19 v0.7.7